api/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
pandas==1.0.3
xlwt==1.3.0
pathlib==1.0.1
nltk==3.4.5
python-dotenv==0.12.0
comtypes==1.1.7
keras==2.4.3
matplotlib==3.3.3
Pillow==2.2.1
seaborn==0.11.0
tensorflow==2.3.1